Separate SearchSpec from SearchConfigSpec
